==> step6_combination.ipynb <==
import pandas as pd
import numpy as np
from lib.utility import get_text, ProcessPipeline
from lib.nlp import nlpPipeline,featurePipeline
import pickle

from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor


# Part 1: data engineering

### read pickle file
with open('outputs/step2_news_raw.pickle', 'rb') as handle:
    texts = pickle.load(handle)

len(texts)

### read pickle file
with open('outputs/step3_LDA_textsProcessed.pickle', 'rb') as handle:
    textsProcessed = pickle.load(handle)

# Part 2: Topic Modeling

from gensim import corpora, models
import gensim
from gensim.models import CoherenceModel


dictionary = corpora.Dictionary(textsProcessed)

dictionary.token2id

corpus = [dictionary.doc2bow(singleDocList) for singleDocList  in textsProcessed]

### This list of tuples represents our first document, doc_a. The tuples are (term ID, term frequency) pairs, so if  print(dictionary.token2id) says brocolli’s id is 0, then the first tuple indicates that brocolli appeared twice in doc_a. doc2bow() only includes terms that actually occur: terms that do not occur in a document will not appear in that document’s vector.
print(corpus[0])

ldamodel = gensim.models.ldamodel.LdaModel(corpus, num_topics=3, id2word = dictionary, passes=20)

ldamodel.print_topics(num_topics=3, num_words=10)

# Visualize the topics
from pyLDAvis import gensim as pyLDAvis_gensim
import pyLDAvis
pyLDAvis.enable_notebook()
vis = pyLDAvis_gensim.prepare(ldamodel, corpus, dictionary)
vis

ldamodel[corpus[0]]

### Check each document and its topics
topicDic = {0:[],1:[],2:[]}
for i in range(len(corpus)):
    allTopics = ldamodel[corpus[i]] ### eg: [(1, 0.58676654), (2, 0.4099236)]
    for topic in allTopics:
        topicID = topic[0]
        prob = topic[1]
        if prob > 0.6:
            topicDic[topicID] += [i]
            break


topicDic.keys()

# Part 3: Sentiment analysis

### read pickle file
with open('outputs/step4_sent_features.pickle', 'rb') as handle:
    sentDF = pickle.load(handle)
print(sentDF.shape)
sentDF.head()

posFeatures = sentDF['pos'].values
negFeatures = sentDF['neg'].values

featureDic = {}
for topicID in topicDic:
    newsIDs = [topicDic[topicID]]
    
    tmpDic = {'pos':posFeatures[newsIDs],'neg':negFeatures[newsIDs]}
    featureDic[topicID] = tmpDic

featureDic

top3Dic = {}
for topicID in featureDic:
    tmpPos = featureDic[topicID]['pos']
    tmpNeg = featureDic[topicID]['neg']
    
    top3pos = np.argpartition(tmpPos, -3)[-3:]
    top3neg =  np.argpartition(tmpNeg, -3)[-3:]
    
    tmpDic = {}
    tmpDic['pos'] = top3pos
    tmpDic['neg'] = top3neg
    
    top3Dic.update({topicID:tmpDic})

top3Dic

for topicID in top3Dic:
    print(f"For topic {topicID}: three most positive articles are : {top3Dic[topicID]['pos']}, three most negative articles are : {top3Dic[topicID]['neg']}")

# Part 4: Keyword extraction

### read pickle file
with open('outputs/step4_sent_textsSummarized.pickle', 'rb') as handle:
    textsSummarized = pickle.load(handle)

entityDic = {}

for topicID in topicDic:
    includedNewsIndex = topicDic[topicID]
    includedNews = np.array(textsSummarized)[includedNewsIndex]
    print("Start to process of topic:",topicID)
    pipeline = nlpPipeline(texts=includedNews,targetPOSs=[]) ### only interested in entity
    entities = pipeline.run()
    entityDic[topicID] = entities
    print("Finish process of topic:",topicID)

entityDic
